Rewrite it in Rust (#15)
* Scrap everything

* First rust commit

Basic Dioxus PWA project. Recreated most of the game but no
functionality yet.

* Ignore compiled tailwind.css

* Restore all screens visually

Still no functionality yet

* Add base funcionality to Game and Help scenes

* Fix styles in Help scene

* Add funcionality to Create and Challange scenes

* Add routing between scenes

* Add animations and visual polish

* Scrap router for custom impl

dioxus_router does not play nice with github pages so replace it with a
cusom impl. It's basically just a match anyway...

* `treman-rs` -> `treman`

This is an artifact from a test run and will break the PWA

* Minor visual fixes